Add utility for simulating a real game
Tametsi is a windows-only game, but I don't want to do dev in a
windows environment. This simulation script uses a screenshot of the
initial board, a screenshot of the final board, and a json file of tile
data to track board state. The `Simulation` object aggregating this
data exposes a `click` function which it can use to update board state,
and a `screenshot` function which will compose an image of the current
board state. By patching this object in for `pyautogui` we can simulate
a full Tametsi game to the solver script without needing to actually
run the game!